<<<<<<< HEAD <<<<<<< HEAD

Channel Conversions

Conversion

Juqst.choi2liou — Function

choi2liou

Takes a superoperator in choi matrix form and returns the liouville superoperator (computational basis)

Juqst.choi2kraus — Function

takes the choi matrix and returns the kraus operators. This stays the same, save that we need to take into account dimensional factors. Note slight negative eigenvalue (rounding can cause problems). Here I round to 15 digits.

Juqst.choiX2liou — Function

takes the Chi matrix (Choi matrix in Pauli basis) and return the liouville superoperator, computaitonal basis

Juqst.choi2chi — Function

Takes a choi matrix and returns the Chi Matrix Note the basis here has changed from marcusps version, The rightmost paulis are varying the quickest. Also note the transpose ( .' ) - this makes it ROW stacking So for instance (0 -im//im 0) vectorises as (0 -im im 0) (unlike column used elsewhere where its (0 im -im 0))

Juqst.kraus2choi — Function

Takes a vector of kraus operators and returns the choi matrix

Juqst.kraus2liou — Function

Takes a vector of kraus operators (matrices) and converts them to a liouville superoperator (computational basis)

Juqst.liou2choi — Function

liou2choi

Take a liouville superoperator (computational basis) and return the choi matrix

Juqst.liou2choiX — Function

takes a liouville superoperator (computational basis) and returns the Chi matrix (basically the choi matrix in Pauli basis)

Juqst.liou2kraus — Function

Takes a liouville superoperator (computational basis) and returns a vector of the equivalen Krawu vectors

Juqst.liou2pauliliou — Function

Converts a liouville superoperator in the computaional basis to one in the pauli basis. Order of Paulis is I,X,Y and Z.

Juqst.pauliliou2liou — Function

Converts a liouville superoperator in the Pauli basis to one in the computational basis. Order of Paulis is I,X,Y and Z.

Manipulation

Juqst.makeSuper — Function

function makeSuper(u)

Takes an operator and turns into a superoperator. Pauli basis.

Juqst.getSuperVec — Function

pass in a density matrix, gives you the super vector that corresponds to it

Verification

Juqst.istp — Function

istp(m;tol)

Returns whether is trace perserving.. Assumes input is in liouville basis (not pauli-liouville - use pauliliou2liou)

Juqst.iscp — Function

iscp(m;tol)

Returns whether is completely positivel. Assumes input is in liouville basis (not pauli-liouville - use pauliliou2liou) Converts to choi and checks the eigenvalues. – tweaked to increase eps slightly was getting a trivial false !cp right on the boundary.

Juqst.ischannel — Function

ischannel(m;tol)

Returns whether is the liouville superoperator is trace perserving and completely positive. Assumes input is in liouville basis (not pauli-liouville - use pauliliou2liou)

Juqst.isunital — Function

unital(m;tol)

Returns whether is the liouville superoperator is unital. Assumes input is in liouville basis (not pauli-liouville - use pauliliou2liou)

Helper

Juqst.genLabels — Function

Helper function for labeling graphs, pass in the qubits get the string of Is and Zs e.g. with two you get ["00","01","10","11"]

Juqst.measure — Method

measure(s,measureNoise, allZs)

Helper function in open-systems.jl Pass in the state, the noise (as a superOperator) and the measurements - use allZs to generate) returns the mapped function (x->x'measureNoises,allZs)

Juqst.genArrows — Function

Helper function (chart axis) return all the ups and downs for a certain number of qubits as a list of strings

Juqst.genZs — Function

returns an array of the 'measurement' operators that extract each of the z measurements e.g. for two qubits will return the II, IZ, ZI and ZZ measurements

=======

Channel Conversions

Conversion

Juqst.choi2liou — Function

choi2liou

Takes a superoperator in choi matrix form and returns the liouville superoperator (computational basis)

Juqst.choi2kraus — Function

takes the choi matrix and returns the kraus operators. This stays the same, save that we need to take into account dimensional factors. Note slight negative eigenvalue (rounding can cause problems). Here I round to 15 digits.

Juqst.choiX2liou — Function

takes the Chi matrix (Choi matrix in Pauli basis) and return the liouville superoperator, computaitonal basis

Juqst.choi2chi — Function

Takes a choi matrix and returns the Chi Matrix Note the basis here has changed from marcusps version, The rightmost paulis are varying the quickest. Also note the transpose ( .' ) - this makes it ROW stacking So for instance (0 -im//im 0) vectorises as (0 -im im 0) (unlike column used elsewhere where its (0 im -im 0))

Juqst.kraus2choi — Function

Takes a vector of kraus operators and returns the choi matrix

Juqst.kraus2liou — Function

Takes a vector of kraus operators (matrices) and converts them to a liouville superoperator (computational basis)

Juqst.liou2choi — Function

liou2choi

Take a liouville superoperator (computational basis) and return the choi matrix

Juqst.liou2choiX — Function

takes a liouville superoperator (computational basis) and returns the Chi matrix (basically the choi matrix in Pauli basis)

Juqst.liou2kraus — Function

Takes a liouville superoperator (computational basis) and returns a vector of the equivalen Krawu vectors

Juqst.liou2pauliliou — Function

Converts a liouville superoperator in the computaional basis to one in the pauli basis. Order of Paulis is I,X,Y and Z.

Juqst.pauliliou2liou — Function

Converts a liouville superoperator in the Pauli basis to one in the computational basis. Order of Paulis is I,X,Y and Z.

Manipulation

Juqst.makeSuper — Function

function makeSuper(u)

Takes an operator and turns into a superoperator. Pauli basis.

Juqst.getSuperVec — Function

pass in a density matrix, gives you the super vector that corresponds to it

Verification

Juqst.istp — Function

istp(m;tol)

Returns whether is trace perserving.. Assumes input is in liouville basis (not pauli-liouville - use pauliliou2liou)

Juqst.iscp — Function

iscp(m;tol)

Returns whether is completely positivel. Assumes input is in liouville basis (not pauli-liouville - use pauliliou2liou) Converts to choi and checks the eigenvalues. – tweaked to increase eps slightly was getting a trivial false !cp right on the boundary.

Juqst.ischannel — Function

ischannel(m;tol)

Returns whether is the liouville superoperator is trace perserving and completely positive. Assumes input is in liouville basis (not pauli-liouville - use pauliliou2liou)

Juqst.isunital — Function

unital(m;tol)

Returns whether is the liouville superoperator is unital. Assumes input is in liouville basis (not pauli-liouville - use pauliliou2liou)

Helper

Missing docstring.

Missing docstring for observeBasis. Check Documenter's build log for details.

Juqst.genLabels — Function

Helper function for labeling graphs, pass in the qubits get the string of Is and Zs e.g. with two you get ["00","01","10","11"]

Juqst.measure — Method

measure(s,measureNoise, allZs)

Helper function in open-systems.jl Pass in the state, the noise (as a superOperator) and the measurements - use allZs to generate) returns the mapped function (x->x'measureNoises,allZs)

Juqst.genArrows — Function

Helper function (chart axis) return all the ups and downs for a certain number of qubits as a list of strings

Juqst.genZs — Function

returns an array of the 'measurement' operators that extract each of the z measurements e.g. for two qubits will return the II, IZ, ZI and ZZ measurements

>>>>>>> 807ee3d (updated) =======

Channel Conversions

Conversion

Juqst.choi2liou — Function

choi2liou

Takes a superoperator in choi matrix form and returns the liouville superoperator (computational basis)

Juqst.choi2kraus — Function

takes the choi matrix and returns the kraus operators. This stays the same, save that we need to take into account dimensional factors. Note slight negative eigenvalue (rounding can cause problems). Here I round to 15 digits.

Juqst.choiX2liou — Function

takes the Chi matrix (Choi matrix in Pauli basis) and return the liouville superoperator, computaitonal basis

Juqst.choi2chi — Function

Takes a choi matrix and returns the Chi Matrix Note the basis here has changed from marcusps version, The rightmost paulis are varying the quickest. Also note the transpose ( .' ) - this makes it ROW stacking So for instance (0 -im//im 0) vectorises as (0 -im im 0) (unlike column used elsewhere where its (0 im -im 0))

Juqst.kraus2choi — Function

Takes a vector of kraus operators and returns the choi matrix

Juqst.kraus2liou — Function

Takes a vector of kraus operators (matrices) and converts them to a liouville superoperator (computational basis)

Juqst.liou2choi — Function

liou2choi

Take a liouville superoperator (computational basis) and return the choi matrix

Juqst.liou2choiX — Function

takes a liouville superoperator (computational basis) and returns the Chi matrix (basically the choi matrix in Pauli basis)

Juqst.liou2kraus — Function

Takes a liouville superoperator (computational basis) and returns a vector of the equivalen Krawu vectors

Juqst.liou2pauliliou — Function

Converts a liouville superoperator in the computaional basis to one in the pauli basis. Order of Paulis is I,X,Y and Z.

Juqst.pauliliou2liou — Function

Converts a liouville superoperator in the Pauli basis to one in the computational basis. Order of Paulis is I,X,Y and Z.

Manipulation

Juqst.makeSuper — Function

function makeSuper(u)

Takes an operator and turns into a superoperator. Pauli basis.

Juqst.getSuperVec — Function

pass in a density matrix, gives you the super vector that corresponds to it

Verification

Juqst.istp — Function

istp(m;tol)

Returns whether is trace perserving.. Assumes input is in liouville basis (not pauli-liouville - use pauliliou2liou)

Juqst.iscp — Function

iscp(m;tol)

Returns whether is completely positivel. Assumes input is in liouville basis (not pauli-liouville - use pauliliou2liou) Converts to choi and checks the eigenvalues. – tweaked to increase eps slightly was getting a trivial false !cp right on the boundary.

Juqst.ischannel — Function

ischannel(m;tol)

Returns whether is the liouville superoperator is trace perserving and completely positive. Assumes input is in liouville basis (not pauli-liouville - use pauliliou2liou)

Juqst.isunital — Function

unital(m;tol)

Returns whether is the liouville superoperator is unital. Assumes input is in liouville basis (not pauli-liouville - use pauliliou2liou)

Helper

Missing docstring.

Missing docstring for observeBasis. Check Documenter's build log for details.

Juqst.genLabels — Function

Helper function for labeling graphs, pass in the qubits get the string of Is and Zs e.g. with two you get ["00","01","10","11"]

Juqst.measure — Method

measure(s,measureNoise, allZs)

Helper function in open-systems.jl Pass in the state, the noise (as a superOperator) and the measurements - use allZs to generate) returns the mapped function (x->x'measureNoises,allZs)

Juqst.genArrows — Function

Helper function (chart axis) return all the ups and downs for a certain number of qubits as a list of strings

Juqst.genZs — Function

returns an array of the 'measurement' operators that extract each of the z measurements e.g. for two qubits will return the II, IZ, ZI and ZZ measurements

=======

Channel Conversions

Conversion

Juqst.choi2liou — Function

choi2liou

Takes a superoperator in choi matrix form and returns the liouville superoperator (computational basis)

Juqst.choi2kraus — Function

takes the choi matrix and returns the kraus operators. This stays the same, save that we need to take into account dimensional factors. Note slight negative eigenvalue (rounding can cause problems). Here I round to 15 digits.

Juqst.choiX2liou — Function

takes the Chi matrix (Choi matrix in Pauli basis) and return the liouville superoperator, computaitonal basis

Juqst.choi2chi — Function

Takes a choi matrix and returns the Chi Matrix Note the basis here has changed from marcusps version, The rightmost paulis are varying the quickest. Also note the transpose ( .' ) - this makes it ROW stacking So for instance (0 -im//im 0) vectorises as (0 -im im 0) (unlike column used elsewhere where its (0 im -im 0))

Juqst.kraus2choi — Function

Takes a vector of kraus operators and returns the choi matrix

Juqst.kraus2liou — Function

Takes a vector of kraus operators (matrices) and converts them to a liouville superoperator (computational basis)

Juqst.liou2choi — Function

liou2choi

Take a liouville superoperator (computational basis) and return the choi matrix

Juqst.liou2choiX — Function

takes a liouville superoperator (computational basis) and returns the Chi matrix (basically the choi matrix in Pauli basis)

Juqst.liou2kraus — Function

Takes a liouville superoperator (computational basis) and returns a vector of the equivalen Krawu vectors

Juqst.liou2pauliliou — Function

Converts a liouville superoperator in the computaional basis to one in the pauli basis. Order of Paulis is I,X,Y and Z.

Juqst.pauliliou2liou — Function

Converts a liouville superoperator in the Pauli basis to one in the computational basis. Order of Paulis is I,X,Y and Z.

Manipulation

Juqst.makeSuper — Function

function makeSuper(u)

Takes an operator and turns into a superoperator. Pauli basis.

Juqst.getSuperVec — Function

pass in a density matrix, gives you the super vector that corresponds to it

Verification

Juqst.istp — Function

istp(m;tol)

Returns whether is trace perserving.. Assumes input is in liouville basis (not pauli-liouville - use pauliliou2liou)

Juqst.iscp — Function

iscp(m;tol)

Returns whether is completely positivel. Assumes input is in liouville basis (not pauli-liouville - use pauliliou2liou) Converts to choi and checks the eigenvalues. – tweaked to increase eps slightly was getting a trivial false !cp right on the boundary.

Juqst.ischannel — Function

ischannel(m;tol)

Returns whether is the liouville superoperator is trace perserving and completely positive. Assumes input is in liouville basis (not pauli-liouville - use pauliliou2liou)

Juqst.isunital — Function

unital(m;tol)

Returns whether is the liouville superoperator is unital. Assumes input is in liouville basis (not pauli-liouville - use pauliliou2liou)

Helper

Juqst.genLabels — Function

Helper function for labeling graphs, pass in the qubits get the string of Is and Zs e.g. with two you get ["00","01","10","11"]

Juqst.measure — Method

Helper function Pass in the state, the noise (superOperator) and the measurements - use allZs to generate) returns the mapped function.

Juqst.genArrows — Function

Helper function (chart axis) return all the ups and downs for a certain number of qubits as a list of strings

Juqst.genZs — Function

returns an array of the 'measurement' operators that extract each of the z measurements e.g. for two qubits will return the II, IZ, ZI and ZZ measurements

>>>>>>> f4aa736 (updated) >>>>>>> b1811be (updated)